LayoutTests/http/tests/security/xss-DENIED-invalid-domain-change-expected.txt

CONSOLE MESSAGE: line 17: SecurityError: The operation is insecure.

PASS: cross-site not access allowed
